requirements.txt
torch>=1.10.0
transformers>=4.20.0
pillow>=9.0.0
ijson>=3.1.0
tqdm>=4.64.0
tensorboard>=2.9.0